deployment/requirements.txt
streamlit
streamlit-option-menu
pymongo
langchain
langchain-openai
langchain-mongodb
python-dotenv
openai